+++ packages/validators/src/index.ts
@@ -1,6 +1,9 @@
 export * from "./airport-conditions";
 export * from "./airport-info";
 export * from "./craft";
+export * from "./error-response";
 export * from "./explanations";
+export * from "./pagination";
 export * from "./plan";
 export * from "./scenario";
+export * from "./scenario-summary";
